refactor: Relocate data layer files to `deprecated` package
This commit moves several files within the `data` module to a new `deprecated` subpackage. This includes DAOs, entities, repositories, and utility classes related to word and word content management.

The following files and their corresponding import statements have been updated:

- `data/anki/*` -> `data/deprecated/anki/*`
- `data/word/*` -> `data/deprecated/word/*`
- `data/wordcontent/*` -> `data/deprecated/wordcontent/*`
- `data/AppDataBase.kt` -> `data/deprecated/AppDataBase.kt`

This change is part of an ongoing effort to refactor and update the data layer.